feat: Implement `SyncLruCache` (#5259)
This is just a proof of concept.

Implement a cache that works from a "&" context.

- get_from_cache_or_insert - gets a value or inserts a computed value
- get_from_cache - tries to get value

TODO:
- [x] rename methods
- [x] add new methods
- [x] benchmarks
- [x] comments
- [x] add more tests.
- [x] one example of replacing real code
- [x] implement example with `lru` (done in another PR)
- [x] Use `once_cell` instead of `lazy_static!` (done in another PR).

See https://github.com/near/nearcore/issues/5145

Was blocked by: https://github.com/near/nearcore/pull/5512
