+++ chain/network/src/lib.rs
@@ -8,6 +8,7 @@ pub use crate::stats::metrics;
 pub use near_network_primitives::types::PeerInfo;
 
 pub(crate) mod common;
+mod network_protocol;
 mod peer;
 mod peer_manager;
 pub mod routing;
